83 - Move book rating to re-usable component
